Backend Code/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.1.1
django-cleanup==9.0.0
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
Markdown==3.7
pillow==10.4.0
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.4.0
